requirements.txt
boto3==1.7.14
botocore==1.10.14
Django==2.0.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
docutils==0.14
jmespath==0.9.3
python-dateutil==2.7.2
pytz==2018.4
s3transfer==0.1.13
six==1.11.0
